python/src/auth/requirements.txt
protobuf==4.21.9
PyYAML==6.0
six==1.16.0
pylint
PyJWT
jedi
Flask
Flask-MySQLdb
